+++ schemas/new-category-schema.ts
@@ -1,0 +1,4 @@
+import { z } from "zod";
+export const newCategorySchema = z.object({
+  name: z.string().min(2, { message: "Category name must be at least 2 characters." }),
+});
