+++ schemas/register-schema.ts
@@ -1,0 +1,7 @@
+import { z } from "zod";
+export const RegisterSchema = z.object({
+  fname: z.string({ message: "Enter a valid first name" }),
+  lname: z.string({ message: "Enter a valid last name" }),
+  email: z.string().email({ message: "Enter a valid email address" }),
+  password: z.string({ message: "Enter a valid password" }).min(8, { message: "password length should be 8 characters" }),
+});
